apps/backend/app/core/config.py
# backend/app/core/config.py
import os
from pydantic_settings import BaseSettings

class Settings(BaseSettings):
    # PostgreSQL Database
    DATABASE_URL: str = os.getenv(
        "DATABASE_URL", 
        "postgresql://postgres:password@localhost/task_track_db"
    )
    API_PREFIX: str = os.getenv("API_PREFIX", "/api")
    JWT_SECRET: str = os.getenv("JWT_SECRET", "devsecret")
    JWT_ALGORITHM: str = os.getenv("JWT_ALGORITHM", "HS256")
    APP_ID: str = os.getenv("APP_ID", "default")

settings = Settings()
