apps/backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
python-jose[cryptography]==3.3.0
pydantic==1.10.12
httpx==0.24.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
#code run command
python -m uvicorn app.main:app --reload